# Conflicts:
#	src/main/java/com/alevel/project/coffee/model/Rating.java

+++ src/main/java/com/alevel/project/coffee/model/Rating.java
@@ -11,7 +11,7 @@ public class Rating implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "rating_id", unique = true, nullable = false)
-    private int id;
+    private Long id;
 
     @Column(name = "ratingFoodQuality")
     private int ratingFoodQuality;
@@ -46,12 +46,11 @@ public class Rating implements Serializable {
         this.place = place;
     }
 
-
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
